Revert "Use double for floating numbers"
This reverts commit 685795bac5fac1cc2a5b15c8a9fbef13596664a9.  The
change does not compile and there is no simple way to fix it without
modifying the AirVantage API.

+++ RouterComponent/router.c
@@ -354,7 +354,7 @@ cleanup:
   return;
 }
 
-void dataRouter_WriteFloat(const char* key, double value, uint32_t timestamp)
+void dataRouter_WriteFloat(const char* key, float value, uint32_t timestamp)
 {
   char appName[SWI_MANGOH_DATA_ROUTER_APP_ID_LEN] = {0};
   le_result_t res = LE_OK;
@@ -583,7 +583,7 @@ cleanup:
   return;
 }
 
-void dataRouter_ReadFloat(const char* key, double* valuePtr, uint32_t* timestampPtr)
+void dataRouter_ReadFloat(const char* key, float* valuePtr, uint32_t* timestampPtr)
 {
   char appName[SWI_MANGOH_DATA_ROUTER_APP_ID_LEN] = {0};
   le_result_t res = LE_OK;
